requirements.txt
python-dotenv==0.19.0
Flask==2.0.1
folium==0.12.1
geopy==2.2.0
requests==2.26.0
